Refactor extracting url values to support more types
