+++ server/service/hosts.go
@@ -2,11 +2,46 @@ package service
 
 import (
 	"context"
+	"time"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 )
+
+// HostResponse is the response struct that contains the full host information
+// along with the host online status and the "display text" to be used when
+// rendering in the UI.
+type HostResponse struct {
+	*fleet.Host
+	Status      fleet.HostStatus `json:"status"`
+	DisplayText string           `json:"display_text"`
+	Labels      []fleet.Label    `json:"labels,omitempty"`
+}
+
+func hostResponseForHost(ctx context.Context, svc fleet.Service, host *fleet.Host) (*HostResponse, error) {
+	return &HostResponse{
+		Host:        host,
+		Status:      host.Status(time.Now()),
+		DisplayText: host.Hostname,
+	}, nil
+}
+
+// HostDetailResponse is the response struct that contains the full host information
+// with the HostDetail details.
+type HostDetailResponse struct {
+	fleet.HostDetail
+	Status      fleet.HostStatus `json:"status"`
+	DisplayText string           `json:"display_text"`
+}
+
+func hostDetailResponseForHost(ctx context.Context, svc fleet.Service, host *fleet.HostDetail) (*HostDetailResponse, error) {
+	return &HostDetailResponse{
+		HostDetail:  *host,
+		Status:      host.Status(time.Now()),
+		DisplayText: host.Hostname,
+	}, nil
+}
 
 ////////////////////////////////////////////////////////////////////////////////
 // List Hosts
@@ -50,7 +85,7 @@ func listHostsEndpoint(ctx context.Context, request interface{}, svc fleet.Servi
 	return listHostsResponse{Hosts: hostResponses, Software: software}, nil
 }
 
-func (svc Service) ListHosts(ctx context.Context, opt fleet.HostListOptions) ([]*fleet.Host, error) {
+func (svc *Service) ListHosts(ctx context.Context, opt fleet.HostListOptions) ([]*fleet.Host, error) {
 	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
 		return nil, err
 	}
@@ -64,7 +99,7 @@ func (svc Service) ListHosts(ctx context.Context, opt fleet.HostListOptions) ([]
 	return svc.ds.ListHosts(ctx, filter, opt)
 }
 
-func (svc Service) SoftwareByID(ctx context.Context, id uint) (*fleet.Software, error) {
+func (svc *Service) SoftwareByID(ctx context.Context, id uint) (*fleet.Software, error) {
 	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
 		return nil, err
 	}
@@ -73,7 +108,7 @@ func (svc Service) SoftwareByID(ctx context.Context, id uint) (*fleet.Software, 
 }
 
 /////////////////////////////////////////////////////////////////////////////////
-// Delete
+// Delete Hosts
 /////////////////////////////////////////////////////////////////////////////////
 
 type deleteHostsRequest struct {
@@ -108,7 +143,7 @@ func deleteHostsEndpoint(ctx context.Context, request interface{}, svc fleet.Ser
 	return deleteHostsResponse{}, nil
 }
 
-func (svc Service) DeleteHosts(ctx context.Context, ids []uint, opts fleet.HostListOptions, lid *uint) error {
+func (svc *Service) DeleteHosts(ctx context.Context, ids []uint, opts fleet.HostListOptions, lid *uint) error {
 	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
 		return err
 	}
@@ -166,7 +201,7 @@ func countHostsEndpoint(ctx context.Context, request interface{}, svc fleet.Serv
 	return countHostsResponse{Count: count}, nil
 }
 
-func (svc Service) CountHosts(ctx context.Context, labelID *uint, opts fleet.HostListOptions) (int, error) {
+func (svc *Service) CountHosts(ctx context.Context, labelID *uint, opts fleet.HostListOptions) (int, error) {
 	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
 		return 0, err
 	}
@@ -174,9 +209,39 @@ func (svc Service) CountHosts(ctx context.Context, labelID *uint, opts fleet.Hos
 	return svc.countHostFromFilters(ctx, labelID, opts)
 }
 
+func (svc *Service) countHostFromFilters(ctx context.Context, labelID *uint, opt fleet.HostListOptions) (int, error) {
+	filter, err := processHostFilters(ctx, opt, nil)
+	if err != nil {
+		return 0, err
+	}
+
+	var count int
+	if labelID != nil {
+		count, err = svc.ds.CountHostsInLabel(ctx, filter, *labelID, opt)
+	} else {
+		count, err = svc.ds.CountHosts(ctx, filter, opt)
+	}
+	if err != nil {
+		return 0, err
+	}
+
+	return count, nil
+}
+
 /////////////////////////////////////////////////////////////////////////////////
 // Get host
 /////////////////////////////////////////////////////////////////////////////////
+
+type getHostRequest struct {
+	ID uint `url:"id"`
+}
+
+type getHostResponse struct {
+	Host *HostDetailResponse `json:"host"`
+	Err  error               `json:"error,omitempty"`
+}
+
+func (r getHostResponse) error() error { return r.Err }
 
 func getHostEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
 	req := request.(*getHostRequest)
@@ -193,7 +258,27 @@ func getHostEndpoint(ctx context.Context, request interface{}, svc fleet.Service
 	return getHostResponse{Host: resp}, nil
 }
 
-func (svc Service) checkWriteForHostIDs(ctx context.Context, ids []uint) error {
+func (svc *Service) GetHost(ctx context.Context, id uint) (*fleet.HostDetail, error) {
+	// First ensure the user has access to list hosts, then check the specific
+	// host once team_id is loaded.
+	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
+		return nil, err
+	}
+
+	host, err := svc.ds.Host(ctx, id, false)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "get host")
+	}
+
+	// Authorize again with team loaded now that we have team_id
+	if err := svc.authz.Authorize(ctx, host, fleet.ActionRead); err != nil {
+		return nil, err
+	}
+
+	return svc.getHostDetails(ctx, host)
+}
+
+func (svc *Service) checkWriteForHostIDs(ctx context.Context, ids []uint) error {
 	for _, id := range ids {
 		host, err := svc.ds.Host(ctx, id, false)
 		if err != nil {
@@ -254,6 +339,127 @@ func (svc *Service) GetHostSummary(ctx context.Context, teamID *uint) (*fleet.Ho
 }
 
 ////////////////////////////////////////////////////////////////////////////////
+// Get Host By Identifier
+////////////////////////////////////////////////////////////////////////////////
+
+type hostByIdentifierRequest struct {
+	Identifier string `url:"identifier"`
+}
+
+func hostByIdentifierEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*hostByIdentifierRequest)
+	host, err := svc.HostByIdentifier(ctx, req.Identifier)
+	if err != nil {
+		return getHostResponse{Err: err}, nil
+	}
+
+	resp, err := hostDetailResponseForHost(ctx, svc, host)
+	if err != nil {
+		return getHostResponse{Err: err}, nil
+	}
+
+	return getHostResponse{
+		Host: resp,
+	}, nil
+}
+
+func (svc *Service) HostByIdentifier(ctx context.Context, identifier string) (*fleet.HostDetail, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
+		return nil, err
+	}
+
+	host, err := svc.ds.HostByIdentifier(ctx, identifier)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "get host by identifier")
+	}
+
+	// Authorize again with team loaded now that we have team_id
+	if err := svc.authz.Authorize(ctx, host, fleet.ActionRead); err != nil {
+		return nil, err
+	}
+
+	return svc.getHostDetails(ctx, host)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Delete Host
+////////////////////////////////////////////////////////////////////////////////
+
+type deleteHostRequest struct {
+	ID uint `url:"id"`
+}
+
+type deleteHostResponse struct {
+	Err error `json:"error,omitempty"`
+}
+
+func (r deleteHostResponse) error() error { return r.Err }
+
+func deleteHostEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*deleteHostRequest)
+	err := svc.DeleteHost(ctx, req.ID)
+	if err != nil {
+		return deleteHostResponse{Err: err}, nil
+	}
+	return deleteHostResponse{}, nil
+}
+
+func (svc *Service) DeleteHost(ctx context.Context, id uint) error {
+	if err := svc.authz.Authorize(ctx, &fleet.Host{}, fleet.ActionList); err != nil {
+		return err
+	}
+
+	host, err := svc.ds.Host(ctx, id, false)
+	if err != nil {
+		return ctxerr.Wrap(ctx, err, "get host for delete")
+	}
+
+	// Authorize again with team loaded now that we have team_id
+	if err := svc.authz.Authorize(ctx, host, fleet.ActionWrite); err != nil {
+		return err
+	}
+
+	return svc.ds.DeleteHost(ctx, id)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Add Hosts to Team
+////////////////////////////////////////////////////////////////////////////////
+
+type addHostsToTeamRequest struct {
+	TeamID  *uint  `json:"team_id"`
+	HostIDs []uint `json:"hosts"`
+}
+
+type addHostsToTeamResponse struct {
+	Err error `json:"error,omitempty"`
+}
+
+func (r addHostsToTeamResponse) error() error { return r.Err }
+
+func addHostsToTeamEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*addHostsToTeamRequest)
+	err := svc.AddHostsToTeam(ctx, req.TeamID, req.HostIDs)
+	if err != nil {
+		return addHostsToTeamResponse{Err: err}, nil
+	}
+
+	return addHostsToTeamResponse{}, err
+}
+
+func (svc *Service) AddHostsToTeam(ctx context.Context, teamID *uint, hostIDs []uint) error {
+	// This is currently treated as a "team write". If we ever give users
+	// besides global admins permissions to modify team hosts, we will need to
+	// check that the user has permissions for both the source and destination
+	// teams.
+	if err := svc.authz.Authorize(ctx, &fleet.Host{TeamID: teamID}, fleet.ActionWrite); err != nil {
+		return err
+	}
+
+	return svc.ds.AddHostsToTeam(ctx, teamID, hostIDs)
+}
+
+////////////////////////////////////////////////////////////////////////////////
 // Add Hosts to Team by Filter
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -288,7 +494,7 @@ func addHostsToTeamByFilterEndpoint(ctx context.Context, request interface{}, sv
 	return addHostsToTeamByFilterResponse{}, err
 }
 
-func (svc Service) AddHostsToTeamByFilter(ctx context.Context, teamID *uint, opt fleet.HostListOptions, lid *uint) error {
+func (svc *Service) AddHostsToTeamByFilter(ctx context.Context, teamID *uint, opt fleet.HostListOptions, lid *uint) error {
 	// This is currently treated as a "team write". If we ever give users
 	// besides global admins permissions to modify team hosts, we will need to
 	// check that the user has permissions for both the source and destination
@@ -351,4 +557,70 @@ func (svc *Service) RefetchHost(ctx context.Context, id uint) error {
 	}
 
 	return nil
+}
+
+func (svc *Service) getHostDetails(ctx context.Context, host *fleet.Host) (*fleet.HostDetail, error) {
+	if err := svc.ds.LoadHostSoftware(ctx, host); err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "load host software")
+	}
+
+	labels, err := svc.ds.ListLabelsForHost(ctx, host.ID)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "get labels for host")
+	}
+
+	packs, err := svc.ds.ListPacksForHost(ctx, host.ID)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "get packs for host")
+	}
+
+	policies, err := svc.ds.ListPoliciesForHost(ctx, host)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "get policies for host")
+	}
+
+	return &fleet.HostDetail{Host: *host, Labels: labels, Packs: packs, Policies: policies}, nil
+}
+
+func (svc *Service) hostIDsFromFilters(ctx context.Context, opt fleet.HostListOptions, lid *uint) ([]uint, error) {
+	filter, err := processHostFilters(ctx, opt, lid)
+	if err != nil {
+		return nil, err
+	}
+
+	// Load hosts, either from label if provided or from all hosts.
+	var hosts []*fleet.Host
+	if lid != nil {
+		hosts, err = svc.ds.ListHostsInLabel(ctx, filter, *lid, opt)
+	} else {
+		hosts, err = svc.ds.ListHosts(ctx, filter, opt)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	if len(hosts) == 0 {
+		return nil, nil
+	}
+
+	hostIDs := make([]uint, 0, len(hosts))
+	for _, h := range hosts {
+		hostIDs = append(hostIDs, h.ID)
+	}
+	return hostIDs, nil
+}
+
+func processHostFilters(ctx context.Context, opt fleet.HostListOptions, lid *uint) (fleet.TeamFilter, error) {
+	vc, ok := viewer.FromContext(ctx)
+	if !ok {
+		return fleet.TeamFilter{}, fleet.ErrNoContext
+	}
+	filter := fleet.TeamFilter{User: vc.User, IncludeObserver: true}
+
+	if opt.StatusFilter != "" && lid != nil {
+		return fleet.TeamFilter{}, fleet.NewInvalidArgumentError("status", "may not be provided with label_id")
+	}
+
+	opt.PerPage = fleet.PerPageUnlimited
+	return filter, nil
 }
