--- server/service/transport_hosts.go
@@ -1,32 +1,0 @@
-package service
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-)
-
-func decodeHostByIdentifierRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	identifier, err := nameFromRequest(r, "identifier")
-	if err != nil {
-		return nil, err
-	}
-	return hostByIdentifierRequest{Identifier: identifier}, nil
-}
-
-func decodeDeleteHostRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	return deleteHostRequest{ID: id}, nil
-}
-
-func decodeAddHostsToTeamRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req addHostsToTeamRequest
-
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, err
-	}
-	return req, nil
-}
