+++ server/service/transport_sessions.go
@@ -11,35 +11,35 @@ import (
 )
 
 func decodeGetInfoAboutSessionRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
+	id, err := uintFromRequest(r, "id")
 	if err != nil {
 		return nil, err
 	}
-	return getInfoAboutSessionRequest{ID: id}, nil
+	return getInfoAboutSessionRequest{ID: uint(id)}, nil
 }
 
 func decodeGetInfoAboutSessionsForUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
+	id, err := uintFromRequest(r, "id")
 	if err != nil {
 		return nil, err
 	}
-	return getInfoAboutSessionsForUserRequest{ID: id}, nil
+	return getInfoAboutSessionsForUserRequest{ID: uint(id)}, nil
 }
 
 func decodeDeleteSessionRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
+	id, err := uintFromRequest(r, "id")
 	if err != nil {
 		return nil, err
 	}
-	return deleteSessionRequest{ID: id}, nil
+	return deleteSessionRequest{ID: uint(id)}, nil
 }
 
 func decodeDeleteSessionsForUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
+	id, err := uintFromRequest(r, "id")
 	if err != nil {
 		return nil, err
 	}
-	return deleteSessionsForUserRequest{ID: id}, nil
+	return deleteSessionsForUserRequest{ID: uint(id)}, nil
 }
 
 func decodeLoginRequest(ctx context.Context, r *http.Request) (interface{}, error) {
